fixed warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@1159 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/RoadFix.cpp
@@ -102,7 +102,7 @@ int RoadMapEdit::MergeRedundantNodes(bool bDegrees, void progress_callback(int))
 			}
 		}
 		count += (nodes - count1);
-		count_tick = count / tick_size;
+		count_tick = (int) (count / tick_size);
 		if (count_tick > count_last_tick)
 		{
 			count_last_tick = count_tick;
